Move database config to EnvironmentConfig
