+++ pkg/portal/db/deps.go
@@ -1,8 +1,8 @@
 package db
 
 import (
+	"github.com/authgear/authgear-server/pkg/lib/config"
 	"github.com/authgear/authgear-server/pkg/lib/infra/db"
-	"github.com/authgear/authgear-server/pkg/portal/config"
 	"github.com/google/wire"
 )
 
@@ -13,7 +13,7 @@ var DependencySet = wire.NewSet(
 	wire.Struct(new(SQLExecutor), "*"),
 )
 
-func NewSQLBuilder(config *config.DatabaseConfig) *db.SQLBuilder {
+func NewSQLBuilder(config *config.DatabaseEnvironmentConfig) *db.SQLBuilder {
 	builder := db.NewSQLBuilder(config.DatabaseSchema, "")
 	return &builder
 }
